requirements.txt
# Core dependencies for XDL Processing
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
astropy>=5.0.0
pyyaml>=6.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
